+++ __tests__/compact.test.ts
@@ -1,7 +1,0 @@
-import { compact } from "../src";
-
-const dummies = ["John", null, false, ""];
-
-test("filters out empty values", () => {
-    expect(compact(dummies)).toEqual(["John"]);
-});