+++ __tests__/sample.test.ts
@@ -1,8 +1,0 @@
-import "jest-extended";
-import { sample } from "../src";
-
-const dummies = ["John", "Jane"];
-
-test("picks a random item", () => {
-    expect(dummies).toContain(sample(dummies));
-});